Create Program.fs
The "REAL" first program. Based on the HTML CANVAS example (by ZachBray).

Tested on a vanilla "FunScript - Not Hosted" project (you can find it through the "online" project option).

1. The whole idea is that you create "code". This is where your main() is used - a.k.a. your magic. Here just copied the "canvas" example.
2. Insert it into an HTML "page". I copied the "canvas" index.html file. And added the script tag into the "body" with reference to the "code" as its content.
3. The rest is standard - putting the "page" into an index.html file and launching your default browser.

I spent 5 hours on this till I got the hang of it. Maybe will be a 5 minute thing for you now :)